refactor: hide `awkward.typing`
