+++ src/awkward/_broadcasting.py
@@ -12,6 +12,7 @@ from awkward._behavior import find_custom_broadcast
 from awkward._nplikes.numpy import Numpy
 from awkward._nplikes.numpylike import NumpyMetadata
 from awkward._nplikes.shape import unknown_length
+from awkward._typing import Any, Callable, Dict, List, TypeAlias, Union
 from awkward._util import unset
 from awkward.contents.bitmaskedarray import BitMaskedArray
 from awkward.contents.bytemaskedarray import ByteMaskedArray
@@ -32,7 +33,6 @@ from awkward.index import (  # IndexU8,  ; Index32,  ; IndexU32,  ; noqa: F401
     Index64,
 )
 from awkward.record import Record
-from awkward.typing import Any, Callable, Dict, List, TypeAlias, Union
 
 np = NumpyMetadata.instance()
 numpy = Numpy.instance()
