+++ src/awkward/_nplikes/shape.py
@@ -1,7 +1,7 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
 from __future__ import annotations
 
-from awkward.typing import Self, TypeAlias
+from awkward._typing import Self, TypeAlias
 
 ShapeItem: TypeAlias = "int | _UnknownLength"
