+++ src/awkward/forms/emptyform.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 import awkward as ak
 from awkward._behavior import find_typestr
 from awkward._errors import deprecate
+from awkward._typing import final
 from awkward._util import unset
 from awkward.forms.form import Form, JSONMapping
-from awkward.typing import final
 
 
 @final
