+++ src/awkward/forms/indexedform.py
@@ -1,9 +1,9 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
 
 import awkward as ak
+from awkward._typing import final
 from awkward._util import unset
 from awkward.forms.form import Form, _parameters_union, _type_parameters_equal
-from awkward.typing import final
 
 
 @final
