+++ src/awkward/forms/regularform.py
@@ -3,9 +3,9 @@ import awkward as ak
 from awkward._behavior import find_typestr
 from awkward._nplikes.shape import unknown_length
 from awkward._regularize import is_integer
+from awkward._typing import final
 from awkward._util import unset
 from awkward.forms.form import Form, _type_parameters_equal
-from awkward.typing import final
 
 
 @final
