+++ src/awkward/operations/ak_from_avro_file.py
@@ -1,7 +1,7 @@
 # BSD 3-Clause License; see https://github.com/scikit-hep/awkward-1.0/blob/main/LICENSE
 __all__ = ("from_avro_file",)
 
-# from awkward.typing import Type
+# from awkward._typing import Type
 import pathlib
 
 import awkward as ak
