+++ src/awkward/types/recordtype.py
@@ -5,9 +5,9 @@ from collections.abc import Iterable
 
 import awkward as ak
 import awkward._prettyprint
+from awkward._typing import final
 from awkward.forms.form import _type_parameters_equal
 from awkward.types.type import Type
-from awkward.typing import final
 
 
 @final
