+++ src/awkward/types/uniontype.py
@@ -3,9 +3,9 @@
 from collections.abc import Iterable
 
 import awkward as ak
+from awkward._typing import final
 from awkward.forms.form import _type_parameters_equal
 from awkward.types.type import Type
-from awkward.typing import final
 
 
 @final
